pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hallucination-risk"
version = "0.1.0"
description = "REST API for hallucination risk assessment using EDFL/B2T/ISR framework"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["LLM", "hallucination", "risk", "EDFL", "OpenAI", "API", "REST"]
authors = [ { name = "Hassana Labs" } ]
license = { text = "MIT" }
urls = { Homepage = "https://hassana.io" }
dependencies = [
  "openai>=1.0.0",
  "flask>=2.3.0",
  "flask-cors>=4.0.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pyinstaller>=6.0.0",
  "pytest>=7.0.0",
  "ruff>=0.4.0",
  "mypy>=1.4.0",
]

[project.scripts]
hallrisk-api = "api.rest_api:main"

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "scripts*"]
